move game service to core layer
